Rename async write method
